Add files via upload
Investors (lenders) provide loans to borrowers in exchange with an agreement of repayment with interest. That means the lender only makes profit (interest) if the borrower pays off the loan. However, if he/she doesn’t repay the loan, then the lender loses money.
XYZ Corp. is looking towrads building a model to predict default in the future, which will help the them in deciding whether sanction a loan or not.